+++ bpm-engine-impl/src/main/java/io/takari/bpm/Configuration.java
@@ -1,0 +1,16 @@
+package io.takari.bpm;
+
+import java.io.Serializable;
+
+public class Configuration implements Serializable {
+
+    private boolean throwExceptionOnErrorEnd = false;
+
+    public boolean isThrowExceptionOnErrorEnd() {
+        return throwExceptionOnErrorEnd;
+    }
+
+    public void setThrowExceptionOnErrorEnd(boolean throwExceptionOnErrorEnd) {
+        this.throwExceptionOnErrorEnd = throwExceptionOnErrorEnd;
+    }
+}